+++ angular_forms/test/integration_test.dart
@@ -39,7 +39,7 @@ void main() {
       });
       expect(
           update,
-          throwsInAngular(allOf(
+          throwsA(allOf(
               new isInstanceOf<Error>(),
               predicate((e) => e.message.contains(
                   'ngFormModel expects a form. Please pass one in.')))));